update for npdocker

+++ Docker/npdocker/clear.sh
@@ -1,15 +1,37 @@
 #!/bin/sh
 target=web
-images=repo.local.com/np:1.1
+ver=1.1
+images=repo.local.com/$target:$ver
 
+echo "clear $target and delete $images(y/n)"
+read key
+if [ $key = 'y' ]; then
 docker stop $target
 docker rm $target
 docker rmi $images
+fi
+
+
 echo "Rebuild for $images(y/n)"
 read key
 if [ $key = 'y' ]; then
 docker build -t $images .
 docker images
-
-
+echo "Mission completed!"
 fi
+
+
+echo "Building a server for $sn?(y/n)"
+read key
+if [ $key = 'y' ]; then
+
+docker run --restart=always -d \
+    --name web \
+    -p 80:80 \
+    -v /html:/html \
+$images
+docker logs $target
+fi
+echo "Visit web http://localhost/ via webrower"
+
+
